Refactor components, now there is only one Factory class
