+++ src/Chronoverse.js
@@ -11,9 +11,9 @@ import ScorePanel from './views/panels/ScorePanel';
 import ButtonsPanelClassic from './views/panels/ButtonsPanelClassic';
 import ButtonsPanelSpaceRace from './views/panels/ButtonsPanelSpaceRace';
 
-import FactoryClassic from './classic/Factory';
-import FactorySpaceRace from './spaceRace/Factory';
+import Factory from './factory';
 import { KEY, GAME_STATE } from './util/constants';
+import { GAME_MODE } from './util/factoryHelper';
 import { getStorageClassicTopScore, getStorageSpaceRaceTopScore, setStorageClassicTopScore, setStorageSpaceRaceTopScore } from './util/localStorageHelper';
 import { PLAYLIST } from './util/soundHelper';
 
@@ -22,7 +22,7 @@ export default class Chronoverse extends Component {
 		super(props);
 		this.actions = props.actions;
 
-		this.appVersion = '1.2.0';
+		this.appVersion = '1.2.5';
 		this.ship = [];
 		this.asteroids = [];
 		this.bullets = [];
@@ -43,8 +43,7 @@ export default class Chronoverse extends Component {
 			setEnemyCount: this.actions.setEnemyCount.bind(this),
 			setPowerUpCount: this.actions.setPowerUpCount.bind(this)
 		};
-		this.factoryClassic = new FactoryClassic(factoryInit);
-		this.factorySpaceRace = new FactorySpaceRace(factoryInit);
+		this.factory = new Factory(factoryInit);
 	}
 
 	getState() {
@@ -84,11 +83,8 @@ export default class Chronoverse extends Component {
 		context.globalAlpha = 1;
 
 		// Next set of elements
-		if (this.getState().game.inClassicGame && !this.asteroids.length) {
-			this.factoryClassic.nextSetOfComponents(this.getState().stats.currentScore, this.getState().enemyCount, this.getState().asteroidCount, this.powerUps);
-		}
-		if (this.getState().game.inSpaceRaceGame && !this.asteroids.length) {
-			this.factorySpaceRace.nextSetOfComponents(this.getState().stats.currentScore, this.getState().enemyCount, this.getState().asteroidCount, this.powerUps);
+		if (this.isInGame() && !this.asteroids.length) {
+			this.factory.nextSetOfComponents(this.getState().stats.currentScore, this.getState().enemyCount, this.getState().asteroidCount, this.powerUps);
 		}
 
 		// Check for colisions
@@ -397,24 +393,26 @@ export default class Chronoverse extends Component {
 
 	startClassicGame() {
 		this.actions.setGameState(GAME_STATE.CLASSIC);
+		this.factory.setGameMode(GAME_MODE.CLASSIC);
 
 		this.asteroids = [];
 		this.bullets = [];
 		this.powerUps = [];
 		this.enemies = [];
 		this.actions.setTopScoreInUse(this.getState().stats.topScoreClassic);
-		this.factoryClassic.generateShip();
+		this.factory.generateShip();
 	}
 
 	startSpaceRaceGame() {
 		this.actions.setGameState(GAME_STATE.SPACE_RACE);
+		this.factory.setGameMode(GAME_MODE.SPACE_RACE);
 
 		this.asteroids = [];
 		this.bullets = [];
 		this.powerUps = [];
 		this.enemies = [];
 		this.actions.setTopScoreInUse(this.getState().stats.topScoreSpaceRace);
-		this.factorySpaceRace.generateShip();
+		this.factory.generateShip();
 	}
 
 	gameOver() {
