+++ src/views/SelectGame.js
@@ -11,7 +11,7 @@ import MdLock from 'react-icons/lib/md/lock'
 import MdInfo from 'react-icons/lib/md/info'
 import FaTrophy from 'react-icons/lib/fa/trophy'
 import { isPassive, isMobileDevice } from '../util/helpers';
-import { GAME_RULES } from '../util/constants';
+import { GAME_MODE } from '../util/factoryHelper';
 import { PLAYLIST } from '../util/soundHelper';
 import { getStorageClassicTopScore, getStorageSpaceRaceTopScore} from '../util/localStorageHelper';
 
@@ -20,10 +20,10 @@ export default class SelectGame extends React.Component {
 		super(props);
 		this.myScroll = null;
 
-		GAME_RULES.CLASSIC.onClickEvent = props.startClassicGame;
-		GAME_RULES.CLASSIC.topScore = getStorageClassicTopScore();
-		GAME_RULES.SPACE_RACE.onClickEvent = props.startSpaceRaceGame;
-		GAME_RULES.SPACE_RACE.topScore = getStorageSpaceRaceTopScore();
+		GAME_MODE.CLASSIC.rules.onClickEvent = props.startClassicGame;
+		GAME_MODE.CLASSIC.rules.topScore = getStorageClassicTopScore();
+		GAME_MODE.SPACE_RACE.rules.onClickEvent = props.startSpaceRaceGame;
+		GAME_MODE.SPACE_RACE.rules.topScore = getStorageSpaceRaceTopScore();
 	}
 
 	componentDidMount() {
@@ -103,11 +103,11 @@ export default class SelectGame extends React.Component {
 				<div id='wrapper' className='wrapper-sg' ref='wrapper'>
 					<div id='scroller' className='scroller-sg'>
 						<ul>
-							{this.getGameSelection(GAME_RULES.CLASSIC,
-								(GAME_RULES.CLASSIC.topScore >= GAME_RULES.CLASSIC.unlockAt))
+							{this.getGameSelection(GAME_MODE.CLASSIC.rules,
+								(GAME_MODE.CLASSIC.rules.topScore >= GAME_MODE.CLASSIC.rules.unlockAt))
 							}
-							{this.getGameSelection(GAME_RULES.SPACE_RACE,
-								(GAME_RULES.CLASSIC.topScore >= GAME_RULES.SPACE_RACE.unlockAt))
+							{this.getGameSelection(GAME_MODE.SPACE_RACE.rules,
+								(GAME_MODE.CLASSIC.rules.topScore >= GAME_MODE.SPACE_RACE.rules.unlockAt))
 							}
 						</ul>
 					</div>
